server/requirements.txt
alembic==1.16.5
amqp==5.3.1
annotated-types==0.7.0
anyio==4.11.0
Authlib==1.6.4
bcrypt==4.3.0
bidict==0.23.1
billiard==4.2.2
blinker==1.9.0
blis==1.3.0
CacheControl==0.14.3
cachelib==0.13.0
cachetools==5.5.2
catalogue==2.0.10
celery==5.5.3
certifi==2025.8.3
cffi==2.0.0
charset-normalizer==3.4.3
click==8.3.0
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
cloudinary==1.44.1
cloudpathlib==0.23.0
colorama==0.4.6
confection==0.1.5
cryptography==46.0.1
cymem==2.0.11
distro==1.9.0
dnspython==2.8.0
en_core_web_sm @ https://github.com/explosion/spacy-models/releases/download/en_core_web_sm-3.8.0/en_core_web_sm-3.8.0-py3-none-any.whl#sha256=1932429db727d4bff3deed6b34cfc05df17794f4a52eeb26cf8928f7c1a0fb85
filelock==3.20.0
firebase_admin==7.1.0
Flask==3.1.2
Flask-Bcrypt==1.0.1
flask-cors==6.0.1
Flask-JWT-Extended==4.7.1
Flask-Mail==0.10.0
Flask-Migrate==4.1.0
Flask-OAuthlib==0.9.6
Flask-SocketIO==5.5.1
Flask-SQLAlchemy==3.1.1
Flask-Limiter==2.1.0
eventlet==0.34.0
fsspec==2025.9.0
google-ai-generativelanguage==0.6.15
google-api-core==2.25.1
google-api-python-client==2.183.0
google-auth==2.40.3
google-auth-httplib2==0.2.0
google-cloud-core==2.4.3
google-cloud-firestore==2.21.0
google-cloud-storage==3.4.1
google-crc32c==1.7.1
google-generativeai==0.8.5
google-resumable-media==2.7.2
googleapis-common-protos==1.70.0
greenlet==3.2.4
grpcio==1.75.0
grpcio-status==1.71.2
gunicorn==23.0.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httplib2==0.31.0
httpx==0.28.1
huggingface-hub==0.35.3
hyperframe==6.1.0
idna==3.10
itsdangerous==2.2.0
Jinja2==3.1.6
jiter==0.11.0
joblib==1.5.2
kombu==5.5.4
langcodes==3.5.0
language_data==1.3.0
Mako==1.3.10
marisa-trie==1.3.1
markdown-it-py==4.0.0
MarkupSafe==3.0.2
mdurl==0.1.2
mpmath==1.3.0
msgpack==1.1.2
murmurhash==1.0.13
networkx==3.5
numpy==2.3.4
oauthlib==2.1.0
openai==1.109.1
packaging==25.0
pillow==11.3.0
preshed==3.0.10
prompt_toolkit==3.0.52
proto-plus==1.26.1
protobuf==5.29.5
psycopg2-binary==2.9.10
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
pydantic==2.11.9
pydantic_core==2.33.2
Pygments==2.19.2
PyJWT==2.10.1
pymongo==4.15.1
PyMuPDF==1.26.4
pyotp==2.9.0
pyparsing==3.2.5
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
python-engineio==4.12.2
python-socketio==5.13.0
PyYAML==6.0.3
qrcode==8.2
redis==6.4.0
regex==2025.9.18
requests==2.32.5
requests-oauthlib==1.1.0
rich==14.2.0
rsa==4.9.1
safetensors==0.6.2
scikit-learn==1.7.2
scipy==1.16.2
sentence-transformers==5.1.1
setuptools==80.9.0
shellingham==1.5.4
simple-websocket==1.1.0
six==1.17.0
smart_open==7.3.1
sniffio==1.3.1
spacy==3.8.7
spacy-legacy==3.0.12
spacy-loggers==1.0.5
SQLAlchemy==2.0.43
srsly==2.5.1
sympy==1.14.0
thinc==8.3.6
threadpoolctl==3.6.0
tokenizers==0.22.1
torch==2.9.0
tqdm==4.67.1
transformers==4.57.1
typer==0.19.2
typing-inspection==0.4.1
typing_extensions==4.15.0
tzdata==2025.2
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wasabi==1.1.3
wcwidth==0.2.14
weasel==0.4.1
Werkzeug==3.1.3
wheel==0.45.1
wrapt==1.17.3
wsproto==1.2.0
bleach==6.0.0